Moves queries to their own folders
